move all block tests into single test suite
